update server dao

+++ src/ReligionServer/src/main/java/com/rss/ReligionServer/service/implementation/SignServiceImpl.java
@@ -3,10 +3,12 @@ package com.rss.ReligionServer.service.implementation;
 import com.rss.ReligionServer.dao.UserDao;
 import com.rss.ReligionServer.model.SignInfo;
 import com.rss.ReligionServer.model.UserModel;
+import com.rss.ReligionServer.model.mapping.UserMapping;
 import com.rss.ReligionServer.service.SignService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -16,10 +18,10 @@ public class SignServiceImpl implements SignService {
 
     @Override
     public int signIn(SignInfo signInfo) {
-        List<UserModel> userModels = userDao.retrieveByEmail(signInfo.getEmail());
+        List<UserMapping> userModels = userDao.retrieveByEmail(signInfo.getEmail());
         if(userModels == null) return 1;
 
-        for(UserModel userModel : userModels) {
+        for(UserMapping userModel : userModels) {
             if(userModel.getPassword().equals(signInfo.getPassword())) return 0;
         }
         return 1;
@@ -27,37 +29,86 @@ public class SignServiceImpl implements SignService {
 
     @Override
     public int create(UserModel userModel) {
-        int userId = (int) userDao.create(userModel);
-        return userId;
+        try {
+            if(userModel == null) throw new Exception();
+
+            UserMapping mapping = new UserMapping(userModel.getId(), userModel.getEmail(), userModel.getPassword(),
+                    userModel.getName(), userModel.getNumber(), userModel.getBirthday(), userModel.getKind());
+
+            int userId = (int) userDao.create(mapping);
+
+            if(userId == -1) throw new Exception();
+
+            return userId;
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return -1;
     }
 
     @Override
     public List<UserModel> findAll() {
-        List<UserModel> userModels = userDao.retrieveAll();
+        List<UserMapping> mappings = userDao.retrieveAll();
+
+        List<UserModel> userModels = new ArrayList<>();
+        for(UserMapping mapping : mappings) {
+            UserModel userModel = new UserModel(mapping.getId(), mapping.getEmail(), mapping.getPassword(), mapping.getName(),
+                    mapping.getNumber(), mapping.getBirthday(), mapping.getKind());
+            userModels.add(userModel);
+        }
         return userModels;
     }
 
     @Override
     public List<UserModel> findUserByEmail(String email) {
-        List<UserModel> userModels = userDao.retrieveByEmail(email);
+        List<UserMapping> mappings = userDao.retrieveByEmail(email);
+
+        List<UserModel> userModels = new ArrayList<>();
+        for(UserMapping mapping : mappings) {
+            UserModel userModel = new UserModel(mapping.getId(), mapping.getEmail(), mapping.getPassword(), mapping.getName(),
+                    mapping.getNumber(), mapping.getBirthday(), mapping.getKind());
+            userModels.add(userModel);
+        }
         return userModels;
     }
 
     @Override
     public UserModel findUserById(int id) {
-        UserModel userModel = userDao.retrieveById(id);
+        UserMapping mapping = userDao.retrieveById(id);
+
+        UserModel userModel = new UserModel(mapping.getId(), mapping.getEmail(), mapping.getPassword(), mapping.getName(),
+                mapping.getNumber(), mapping.getBirthday(), mapping.getKind());
+
         return userModel;
     }
 
     @Override
     public int modifyUser(UserModel userModel) {
-        boolean isUpdate = userDao.update(userModel) > 0;
-        return isUpdate ? 0 : 1;
+        try {
+            if(userModel == null) throw new Exception();
+            UserMapping mapping = new UserMapping(userModel.getId(), userModel.getEmail(), userModel.getPassword(),
+                    userModel.getName(), userModel.getNumber(), userModel.getBirthday(), userModel.getKind());
+
+            boolean isUpdate = userDao.update(mapping) > 0;
+
+            return isUpdate ? 1 : 0;
+        } catch(Exception e) {
+            e.printStackTrace();
+        }
+        return 1;
     }
 
     @Override
     public int deleteUser(int id) {
-        boolean isDelete = userDao.remove(id) > 0;
-        return isDelete ? 0 : 1;
+        try {
+            UserMapping mapping = userDao.retrieveById(id);
+            if(mapping == null) throw new Exception();
+
+            boolean isDelete = userDao.remove(id) > 0;
+            return isDelete ? 0 : 1;
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return 1;
     }
 }
